test_aassaasa/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --config=pyproject.toml requirements/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.7.2
    # via
    #   channels
    #   channels-redis
    #   django
async-timeout==4.0.3
    # via redis
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
billiard==4.1.0
    # via celery
boto3==1.28.38
    # via -r requirements/requirements.in
botocore==1.31.38
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.3.4
    # via -r requirements/requirements.in
celery-progress==0.3
    # via -r requirements/requirements.in
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
channels==4.0.0
    # via
    #   -r requirements/requirements.in
    #   channels-redis
channels-redis==4.1.0
    # via -r requirements/requirements.in
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==41.0.3
    # via
    #   django-anymail
    #   pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-stripe==2.8.1
    # via -r requirements/requirements.in
django==4.2.9
    # via
    #   -r requirements/requirements.in
    #   channels
    #   dj-stripe
    #   django-allauth
    #   django-allauth-2fa
    #   django-anymail
    #   django-hijack
    #   django-otp
    #   django-storages
    #   django-waffle
    #   djangorestframework
    #   drf-spectacular
django-allauth==0.57.0
    # via
    #   -r requirements/requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.11.1
    # via -r requirements/requirements.in
django-anymail[mailgun]==10.1
    # via -r requirements/requirements.in
django-environ==0.11.2
    # via -r requirements/requirements.in
django-hijack==3.4.1
    # via -r requirements/requirements.in
django-otp==1.2.2
    # via django-allauth-2fa
django-storages==1.13.2
    # via -r requirements/requirements.in
django-waffle==4.0.0
    # via -r requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
djangorestframework-api-key==2.3.0
    # via -r requirements/requirements.in
docutils==0.20.1
    # via -r requirements/requirements.in
drf-spectacular==0.26.4
    # via -r requirements/requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.19.0
    # via drf-spectacular
jsonschema-specifications==2023.7.1
    # via jsonschema
kombu==5.3.2
    # via celery
msgpack==1.0.6
    # via channels-redis
oauthlib==3.2.2
    # via requests-oauthlib
prompt-toolkit==3.0.39
    # via click-repl
psycopg2-binary==2.9.7
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via django-allauth
pypng==0.20220715.0
    # via qrcode
python-dateutil==2.8.2
    # via
    #   botocore
    #   celery
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
qrcode==7.4.2
    # via django-allauth-2fa
redis==4.6.0
    # via
    #   celery
    #   channels-redis
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
requests-oauthlib==1.3.1
    # via django-allauth
rpds-py==0.10.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.6.2
    # via boto3
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
stripe==4.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
typing-extensions==4.7.1
    # via qrcode
tzdata==2023.3
    # via celery
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.16
    # via
    #   botocore
    #   django-anymail
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
